keep style hook contained in individual file
